requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==6.0.1
idna==3.4
packaging==23.0
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
pywin32==305
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
websocket-client==1.5.1

